requirements.txt
streamlit>=1.42.0
google-genai>=1.0.0
openai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
typecast-python>=0.1.5
requests>=2.28.0
